requirements.txt
# UI and App
gradio>=3.40.0
Pillow==10.3.0  # for image display in Streamlit

# Embedding Models
sentence-transformers==2.5.1  # for models like all-MiniLM-L6-v2
openai>=1.84.0           # if using OpenAI embeddings

# Vector Database
faiss-cpu

# Optional: Qwen and HuggingFace support
transformers==4.41.2
torch==2.3.0
accelerate==0.30.1

# Data and Utilities
pandas==2.3.0
numpy==1.26.4
pdfplumber==0.11.6
python-docx==1.1.2
docx==0.2.4
tqdm==4.66.4

# Optional dependencies for LangChain integrations
langchain==0.3.25
langchain-openai==0.3.21
langchain-google-genai==2.1.5
langchain-groq==0.3.2
pydantic>=2.8.2

# Other utilities
anyio==4.3.0
requests==2.32.3
python-dotenv~=1.1.0